pyproject.toml
[tool.poetry]
name = "loans-manager"
version = "0.1.0"
description = ""
authors = ["Daniel Gil <jodagito@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
markdown = "^3.7"
pandas = "^2.2.3"
reportlab = "^4.2.5"
drf-yasg = "^1.21.8"
drf-nested-routers = "^0.94.1"
xlsxwriter = "^3.2.1"
django-extensions = "^3.2.3"
django-cors-headers = "^4.7.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
whitenoise = "^6.9.0"
django-environ = "^0.12.0"
pytest-django = "^4.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
